requirements.txt
streamlit>=1.24.0
pandas>=2.2.3
openai>=1.12.0
fpdf2>=2.7.8
pillow>=10.2.0
numpy>=1.24.0
python-docx>=1.1.0
gspread>=6.0.2
oauth2client>=4.1.3
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
openpyxl>=3.1.2
scikit-learn>=1.4.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-dotenv>=1.0.0


